Update radionuclide source to rely on base source
Change the egs_radionuclude_source to take an input "base source"
instead of defining a collimated or isotropic source. This way, any of
the egs++ sources or a custom source can now be used to define the
distribution of the radionuclide. Also, add warning messages for
deprecated radionuclide source inputs. This change was originally
motivated by simulations of targeted radionuclide therapy.