EGS_Mesh: add mevegs application based on tutor7pp
We'd like to write an output file based on the mesh when the simulation
is done. This is most convenient to do in something similar to
Mevegs_Application::outputResults but requires specific knowledge of the
EGS_Mesh geometry (e.g., region volumes). We dynamic_cast the base
geometry to a EGS_Mesh pointer to access its data and prepare to save a
Gmsh file.

Also, copy the input file to a new output file and output result data
vectors. We're writing out scoring data to Gmsh for quick inspection
after a simulation.

Increase the maximum number of media to 200 for ICRP 145: the MRCP_AF
and AM models have 52 media each, and ICRP 145 meshes have ~190 organs
(media) and are both around 8M elements.