EGS_Mesh: perf: reorder mesh by distance to source
Implement mesh reordering to try and reduce time spent in the isWhere
loop. Simulation results are unchanged (as they should be). A first
measurement is promising: a reduction from 75% to 5% of simulation time
spent in isWhere. There is a noticeable difference when running
simulations too: time spent in one batch went from ~13s to 4s.

The EGS_Vector used as the reference point is hard-coded for the current
test. Later commits will instead use the position of a particle
generated by the source.