EGS_Mesh: fix stuck particles at tet corners
Before this change, simple two-element infinite loops were detected, but
not those involving more elements. This change tracks small steps and if
a potential stuck particle is detected, pushes the particle along the
momentum vector to continue the simulation. This case was detected with
an internal point source simulation where the point source was exactly
on a node shared by many tetrahedra.