Add script to check authors in file and in commits
Authors of the EGSnrc files are unequivocally identified in the git
commit log. However, we also want to preserve the names of authors and
contributors inside text files headers, as much as possible, in case the
files are extracted from the repository (or the repository eventually
moves to another version control system).

This script compares, for each modified file in a commit range (passed
as the first argument, master..HEAD by default), the commit authors of
the file against the authors and contributors listed in the file header.
It reports the missing names, and returns 1 (error) if some authors are
missing, or 0 (success) otherwise. Binary files are excluded from the
comparison, as well as files that do not contain authorship information.
The latter are reported at the end of the output, skipping over files
that do not contain author information (e.g., image files).

This script is intended to run automatically as a github action. It
implies that all commit authors should be listed under "Contributors",
regardless of the significance of the contribution, so that the script
runs cleanly. The output of the github action should always be inspected
before merging to check the files that don't contain authorship
information.