Add script to check authors in file and in commits
Since moving to git (about 2015), authors of changes in the EGSnrc
project are unequivocally identified in the git commit log. However, we
also preserve the names of authors and contributors inside headers of
source files as much as possible, in case the files are extracted from
the repository (or the repository eventually moves to another version
control system). It also allows us to preserve historical authorship, as
well as that for contributions not committed by the author.

For each modified file in a commit range (passed as the first argument,
master..HEAD by default), the script compares the commit authors of the
file against the authors and contributors listed in the file header. It
reports the missing names, and returns 1 (error) if some authors are
missing, or 0 (success) otherwise. Binary files are excluded from the
comparison, as well as files that do not contain authorship information.
The latter are reported at the end of the output.

This script is intended to run automatically as a github action. It
implies that all commit authors should be listed under "Contributors",
regardless of the significance of the contribution, so that the script
runs cleanly. One exception is that author Tessier is excludedi for
convenience, since I typically manage formatting-only and merge commits.
The output of the github action should always be inspected before
merging to review files that don't contain authorship information.