Change radionuclide spectrum to be part of radionuclide source
Change the radionuclide spectrum so that it is no longer part of
egs_spectra.cpp. Instead, it is in egs_radionuclide_source, because it
is not compatible with any other source. Also, add virtual functions to
the base source class to be able to extract information from the
radionuclide source without needing to do a static cast. These functions
were also added to the source collection and transformed source classes.